✨ Feat : 메타데이터 컴포넌트화 및 링크 이미지 적용
링크는 현재페이지 이미지는 유뷰트 썸네일이 들어가게 적용
